// src/Finanza.Client/src/app/core/models/category.model.ts
export interface CategoryTotal {
  received: number;
  spent: number;
  balance: number;
}

export interface Category {
  id: string;
  name: string;
  description?: string;
  total: CategoryTotal;
}

export interface CreateCategoryRequest {
  name: string;
  description?: string;
}

export interface UpdateCategoryRequest {
  name: string;
  description?: string;
}